Create js assessment
this is the code about a variable which holds the the number of NFT's after creation when the metadata is passed to the function as parameters and when the NFT is ready it is stored in the variable and then with the help of another function the total no. of NFT is printed to the console.